Convert the parameters for Image.resize from TileGrid to integers

+++ displayio/tilegrid.py
@@ -275,8 +275,10 @@ class TileGrid:
             if absolute_transform.scale > 1:
                 image = image.resize(
                     (
-                        pixel_width * absolute_transform.scale,
-                        pixel_height * absolute_transform.scale,
+                        int(pixel_width * absolute_transform.scale),
+                        int(
+                            pixel_height * absolute_transform.scale,
+                        ),
                     ),
                     resample=Image.NEAREST,
                 )
